requirements.txt
Django>=4.2
gunicorn
pdfplumber
python-docx
scikit-learn
